fix: remove fix-permissions run on CONDA_DIR
Add additional `INSTALLER` as environment variable to use mamba/conda.
Add `initial-condarc` to set up conda configuration.

We could use an if statement to check if the image uses mamba or conda,
and then install required packages, but this would introduce a lot more
complexity to the Dockerfile. Instead, we could trade some disk space
for simplicity.